+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.30.0
 
-- **FEAT**: Add `ShadKeyboardToolbar` component to show a toolbar above the keyboard.
+- **FEAT**: Add `ShadKeyboardToolbar` and `ShadDefaultKeyboardToolbar` components to show a toolbar above the keyboard. Add `keyboardToolbarBuilder` to `ShadInput`, `ShadInputFormField`, `ShadTextArea`, `ShadTextAreaFormField`, `ShadInputOTP`, `ShadInputOTPFormField` to easily add a keyboard toolbar to these components.
 
 ## 0.29.4